\name{RCassandraDriver-package}
\alias{RCassandraDriver-package}
\alias{RCassandraDriver}
\docType{package}
\title{
\packageTitle{RCassandraDriver}
}
\description{
\packageDescription{RCassandraDriver}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RCassandraDriver}
\packageIndices{RCassandraDriver}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RCassandraDriver}

Maintainer: \packageMaintainer{RCassandraDriver}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
